post_service/requirements.txt
grpcio==1.70.0
grpcio-tools==1.70.0
psycopg2-binary==2.9.1
SQLAlchemy==1.4.23
pytest==7.3.1
pytest-asyncio==0.21.0
pytest-cov==4.1.0
